mautic/scripts/requirements.txt
openai>=1.6.0
Flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
seaborn>=0.13.0
matplotlib>=3.8.0
